+++ build.sh
@@ -3,10 +3,10 @@
 set -xe
 
 CC="${CXX:-cc}"
-PKGS="sdl2 glew freetype2"
+PKGS="sdl2 SDL2_image glew freetype2"
 CFLAGS="-Wall -Wextra -std=c11 -pedantic -ggdb"
 LIBS=-lm
-SRC="src/main.c src/app.c src/la.c src/editor.c src/file_browser.c src/free_glyph.c src/simple_renderer.c src/common.c src/lexer.c src/vim.c src/command.c src/config.c"
+SRC="src/main.c src/app.c src/la.c src/editor.c src/file_browser.c src/free_glyph.c src/simple_renderer.c src/common.c src/lexer.c src/vim.c src/command.c src/config.c src/image.c"
 
 if [ `uname` = "Darwin" ]; then
     CFLAGS+=" -framework OpenGL"